+++ rioxarray/raster_dataset.py
@@ -312,15 +312,18 @@ class RasterDataset(XRasterBase):
     def to_raster(
         self,
         raster_path,
-        driver="GTiff",
+        driver=None,
         dtype=None,
         tags=None,
         windowed=False,
         recalc_transform=True,
+        lock=None,
         **profile_kwargs,
     ):
         """
         Export the Dataset to a raster file. Only works with 2D data.
+
+        ..versionadded:: 0.2 lock
 
         Parameters
         ----------
@@ -328,16 +331,19 @@ class RasterDataset(XRasterBase):
             The path to output the raster to.
         driver: str, optional
             The name of the GDAL/rasterio driver to use to export the raster.
-            Default is "GTiff".
+            Default is "GTiff" if rasterio < 1.2 otherwise it will autodetect.
         dtype: str, optional
             The data type to write the raster to. Default is the datasets dtype.
         tags: dict, optional
             A dictionary of tags to write to the raster.
         windowed: bool, optional
             If True, it will write using the windows of the output raster.
-            This only works if the output raster is tiled. As such, if you
-            set this to True, the output raster will be tiled.
+            This is useful for loading data in chunks when writing. Does not
+            do anything when writing with dask.
             Default is False.
+        lock: boolean or Lock, optional
+            Lock to use to write data using dask.
+            If not supplied, it will use a single process for writing.
         **profile_kwargs
             Additional keyword arguments to pass into writing the raster. The
             nodata, transform, crs, count, width, and height attributes
@@ -379,5 +385,6 @@ class RasterDataset(XRasterBase):
             tags=tags,
             windowed=windowed,
             recalc_transform=recalc_transform,
+            lock=lock,
             **profile_kwargs,
         )