+++ docs/history.rst
@@ -6,8 +6,11 @@ Latest
 - ENH: Added `rio.estimate_utm_crs` (issue #181)
 - ENH: Add support for merging datasets with different CRS (issue #173)
 - ENH: Add support for using dask in `rio.to_raster` (issue #9)
+- ENH: Use the list version of `transform_geom` with rasterio 1.2+ (issue #180)
+- ENH: Support driver autodetection with rasterio 1.2+ (issue #180)
 - BUG: Allow `rio.write_crs` when spatial dimensions not found (pull #186)
 - ENH: Allow multithreaded, lockless reads with `rio.read_xarary` (issue #214)
+- BUG: Update to support rasterio 1.2+ merge (issue #180)
 
 0.1.1
 ------